+++ sesrv-script.bash
@@ -79,7 +79,7 @@ APPID="298740"
 WINE_ARCH="win64"
 WINE_PREFIX_GAME_DIR="drive_c/Games/SpaceEngineersDedicated" #Server executable directory
 WINE_PREFIX_GAME_EXE="DedicatedServer64/SpaceEngineersDedicated.exe -console" #Server executable
-WINE_PREFIX_GAME_CONFIG="drive_c/Games/SpaceEngineersDedicated" #Server save and configuration location
+WINE_PREFIX_GAME_CONFIG="drive_c/users/$USER/Application Data/SpaceEngineersDedicated" #Server save and configuration location
 
 #Ramdisk configuration
 TMPFS_ENABLE=$(cat $SCRIPT_DIR/$SERVICE_NAME-config.conf | grep tmpfs_enable | cut -d = -f2) #Get configuration for tmpfs
@@ -1521,8 +1521,9 @@ script_install_prefix() {
 			fi
 			read -p "Do you want to keep the game installation and server data (saves,configs,etc.)? (y/n): " REINSTALL_PREFIX_KEEP_DATA
 			if [[ "$REINSTALL_PREFIX_KEEP_DATA" =~ ^([yY][eE][sS]|[yY])$ ]]; then
-				mkdir -p $BCKP_DIR/prefix_backup/game
+				mkdir -p $BCKP_DIR/prefix_backup/{game,appdata}
 				mv "$SRV_DIR/$WINE_PREFIX_GAME_DIR"/* $BCKP_DIR/prefix_backup/game
+				mv "$SRV_DIR/$WINE_PREFIX_GAME_CONFIG"/* $BCKP_DIR/prefix_backup/appdata
 			fi
 			rm -rf $SRV_DIR
 			Xvfb :5 -screen 0 1024x768x16 &
@@ -1530,12 +1531,14 @@ script_install_prefix() {
 			env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks corefonts
 			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q vcrun2013
 			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q vcrun2017
-			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet472
+			env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet48
 			env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks sound=disabled
 			pkill -f Xvfb
 			if [[ "$REINSTALL_PREFIX_KEEP_DATA" =~ ^([yY][eE][sS]|[yY])$ ]]; then
 				mkdir -p "$SRV_DIR/$WINE_PREFIX_GAME_DIR"
+				mkdir -p "$SRV_DIR/$WINE_PREFIX_GAME_CONFIG"
 				mv $BCKP_DIR/prefix_backup/game/* "$SRV_DIR/$WINE_PREFIX_GAME_DIR"
+				mv $BCKP_DIR/prefix_backup/appdata/* "$SRV_DIR/$WINE_PREFIX_GAME_CONFIG"
 				rm -rf $BCKP_DIR/prefix_backup
 			fi
 			echo "$(date +"%Y-%m-%d %H:%M:%S") [$VERSION] [$NAME] [INFO] (Reinstall Wine prefix) Wine prefix reinstallation complete." | tee -a "$LOG_SCRIPT"
@@ -1778,6 +1781,12 @@ script_diagnostics() {
 		echo "Notification sending service present: No"
 	fi
 	
+	if [ -f "$SRV_DIR/$WINE_PREFIX_GAME_DIR/DedicatedServer64/SpaceEngineersDedicated.exe" ]; then
+		echo "Game executable present: Yes"
+	else
+		echo "Game executable present: No"
+	fi
+	
 	echo "Diagnostics complete."
 }
 
@@ -1806,7 +1815,11 @@ script_install_packages() {
 			
 			if [[ "$UBUNTU_CODENAME" == "bionic" || "$UBUNTU_CODENAME" == "eoan" || "$UBUNTU_CODENAME" == "focal" ]]; then
 				#Add i386 architecture support
+				apt install --yes sudo gnupg
 				sudo dpkg --add-architecture i386
+				
+				#Install software properties common
+				sudo apt install --yes software-properties-common
 				
 				#Check codename and install config for installation
 				if [[ "$UBUNTU_CODENAME" == "bionic" ]]; then
@@ -1843,7 +1856,7 @@ script_install_packages() {
 				sudo systemctl start smbd nmbd winbind
 				
 				#Install winetricks
-				wget  https://raw.githubusercontent.com/Winetricks/winetricks/master/src/winetricks
+				wget https://raw.githubusercontent.com/Winetricks/winetricks/master/src/winetricks
 				sudo mv winetricks /usr/local/bin/
 				sudo chmod +x /usr/local/bin/winetricks
 			else
@@ -1851,6 +1864,65 @@ script_install_packages() {
 				echo "Exiting"
 				exit 1
 			fi
+		elif [[ "$DISTRO" == "debian" ]]; then
+			#Debian distro
+			
+			#Get codename
+			DEBIAN_CODENAME=$(cat /etc/os-release | grep "^VERSION_CODENAME=" | cut -d = -f2)
+			
+			if [[ "$DEBIAN_CODENAME" == "buster" ]]; then
+				#Add i386 architecture support
+				apt install --yes sudo gnupg
+				sudo dpkg --add-architecture i386
+				
+				#Install software properties common
+				sudo apt install --yes software-properties-common
+				
+				#Add non-free repo for steamcmd
+				sudo apt-add-repository non-free
+				
+				#Check codename and install backport repo if needed
+				if [[ "$DEBIAN_CODENAME" == "buster" ]]; then
+					sudo apt-add-repository "deb http://deb.debian.org/debian $DEBIAN_CODENAME-backports main"
+					sudo apt update
+					sudo apt -t buster-backports install --yes "tmux"
+				fi
+			
+				#Add wine and libfaudio0 repositroy and install packages
+				wget -nc https://dl.winehq.org/wine-builds/winehq.key
+				sudo apt-key add winehq.key
+				
+				wget -nc https://download.opensuse.org/repositories/Emulators:/Wine:/Debian/Debian_10/Release.key
+				sudo apt-key add Release.key
+				
+				sudo apt-add-repository "deb https://dl.winehq.org/wine-builds/debian/ $DEBIAN_CODENAME main"
+				sudo apt-add-repository "deb https://download.opensuse.org/repositories/Emulators:/Wine:/Debian/Debian_10 ./"
+				
+				#Check for updates and update local repo database
+				sudo apt update
+				
+				#Install packages and enable services
+				sudo apt install --yes libfaudio0:i386
+				sudo apt install --yes libfaudio0
+				sudo apt install --yes --install-recommends winehq-staging
+				sudo apt install --yes --install-recommends steamcmd
+				sudo apt install --yes rsync cabextract unzip p7zip wget curl postfix zip jq xvfb samba winbind
+				sudo systemctl enable smbd nmbd winbind
+				sudo systemctl start smbd nmbd winbind
+				
+				#Install winetricks
+				wget https://raw.githubusercontent.com/Winetricks/winetricks/master/src/winetricks
+				sudo mv winetricks /usr/local/bin/
+				sudo chmod +x /usr/local/bin/winetricks
+			else
+				echo "Error: This version of Debian is not supported. Supported versions are: Debian 10 (Buster)"
+				echo "Exiting"
+				exit 1
+			fi
+		else
+			echo "Error: This distro is not supported. This script currently supports Arch Linux, Ubuntu 18.04 LTS (Bionic Beaver), Ubuntu 19.10 (Disco Dingo), Ubuntu 20.04 LTS (Focal Fossa), Debian 10 (Buster). If you want to try the script on your distro, install the packages manually. Check the readme for required package versions."
+			echo "Exiting"
+			exit 1
 		fi
 			
 		if [[ "$DISTRO" == "arch" ]]; then
@@ -1859,7 +1931,7 @@ script_install_packages() {
 		echo "Package installation complete."
 	else
 		echo "os-release file not found. Is this distro supported?"
-		echo "This script currently supports Arch Linux, Ubuntu 18.04 LTS (Bionic Beaver), Ubuntu 19.10 (Disco Dingo), Ubuntu 20.04 LTS (Focal Fossa)"
+		echo "This script currently supports Arch Linux, Ubuntu 18.04 LTS (Bionic Beaver), Ubuntu 19.10 (Disco Dingo), Ubuntu 20.04 LTS (Focal Fossa), Debian 10 (Buster)"
 		exit 1
 	fi
 }
@@ -2091,11 +2163,6 @@ script_install() {
 	echo "Configuration complete. Begining installation..."
 	sleep 3
 	
-	echo "Enabling linger"
-	sudo mkdir -p /var/lib/systemd/linger/
-	sudo touch /var/lib/systemd/linger/$USER
-	sudo mkdir -p /home/$USER/.config/systemd/user
-	
 	echo "Installing bash profile"
 	cat >> /home/$USER/.bash_profile <<- 'EOF'
 	#
@@ -2118,6 +2185,16 @@ script_install() {
 	chmod +x $SCRIPT_DIR/$SCRIPT_NAME
 	touch $SCRIPT_DIR/$SERVICE_NAME-server-list.txt
 	sudo chown -R $USER:users /home/$USER
+	
+	echo "Enabling linger"
+	
+	sudo loginctl enable-linger $USER
+	
+	if [ -d /var/lib/systemd/linger/$USER ]; then
+		sudo mkdir -p /var/lib/systemd/linger/
+		sudo touch /var/lib/systemd/linger/$USER
+		sudo mkdir -p /home/$USER/.config/systemd/user
+	fi
 	
 	echo "Enabling services"
 	
@@ -2175,7 +2252,7 @@ script_install() {
 	env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks corefonts
 	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q vcrun2013
 	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q vcrun2017
-	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet472
+	env DISPLAY=:5.0 WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks -q dotnet48
 	env WINEARCH=$WINE_ARCH WINEDEBUG=-all WINEPREFIX=$SRV_DIR winetricks sound=disabled
 	
 	pkill -f Xvfb
@@ -2213,8 +2290,6 @@ script_install() {
 	
 	echo "Installation complete"
 	echo ""
-	echo "Copy your SSK.txt to $BCKP_SRC_DIR"
-	echo "After you copied your SSK.txt reboot the server and the game server will start on boot."
 	echo "You can login to your $USER account with <sudo -i -u $USER> from your primary account or root account."
 	echo "The script was automaticly copied to the scripts folder located at $SCRIPT_DIR"
 	echo "For any settings you'll want to change, edit the $SCRIPT_DIR/$SERVICE_NAME-config.conf file."